==> Algorithm/4_impute_new_data.ipynb <==
# Preprocessing

## Objetivos del Cuaderno

1. **Manejo de Valores Faltantes** (KNNImputer o SoftImpute o Dror de filas con valores faltantes)

Imports

import pandas as pd
from sklearn.impute import KNNImputer
from fancyimpute import SoftImpute
import warnings

warnings.filterwarnings("ignore")

df_features_final = pd.read_csv("dataset/generated_features.csv", index_col='cups')

### Correcion de errores y NaN

**Técnicas Implementadas**

### 1. KNNImputer

* **Concepto:** Imputación basada en los $k$-vecinos más cercanos ($k=5$).
* **Ventajas:**
    * Mantiene relaciones no lineales entre variables.
    * Considera la similitud entre registros completos.

### 2. SoftImpute

* **Concepto:** Imputación basada en factorización matricial (descomposición SVD).
* **Ventajas:**
    * Eficiente para matrices grandes y dispersas.
    * Maneja bien patrones de datos faltantes aleatorios.


# Opción 1: KNNImputer
knn_imputer = KNNImputer(n_neighbors=5)
df_knn = pd.DataFrame(
    knn_imputer.fit_transform(df_features_final),
    columns=df_features_final.columns,
    index=df_features_final.index
)
df_knn.to_csv("dataset/features_KNNImputer.csv")

# Opción 2: SoftImpute
soft_imputer = SoftImpute()
df_soft = pd.DataFrame(
    soft_imputer.fit_transform(df_features_final),
    columns=df_features_final.columns,
    index=df_features_final.index
)
df_soft.to_csv("dataset/features_SoftImpute.csv")


# Opción 3: Dataset original eliminando filas con NaN
df_drop = df_features_final.copy()

# Voy a eliminar dos filas (cups) que tienen NaN durante todo el invierno
cups_con_nan_invierno = ['ceaddbf817fc', 'd0fbcc1108d8']
df_drop.drop(index=cups_con_nan_invierno, inplace=True, errors='ignore')

# Voy a eliminar tres filas (cups) que tienen NaN durante todo el otoño
cups_con_nan_otoño = ['83c7fbada9b4', 'b476034a2e3d', 'ba38270a360e']
df_drop.drop(index=cups_con_nan_otoño, inplace=True, errors='ignore')

df_drop.to_csv("dataset/features_drop.csv")

tiene_nans = df_knn.isnull().any().any()

if tiene_nans:

    columnas_con_nans = df_features_final.columns[df_features_final.isnull().any()].tolist()
    print(f"Las columnas con valores NaN son: {columnas_con_nans}")

    nans_por_columna = df_features_final.isnull().sum()
    print("\nNúmero de NaNs por columna:")
    print(nans_por_columna[nans_por_columna > 0])

else:
    print("El dataset no contiene ningún valor NaN.")
