==> Algorithm/9_unify.ipynb <==
import pandas as pd

df_kmeans = pd.read_csv('dataset/cluster_cups_kmeans.csv')
df_bkmeans = pd.read_csv('dataset/cluster_cups_bkmeans.csv')
df_sc = pd.read_csv('dataset/cluster_cups_sc.csv')

# Renombrar las columnas para identificar cada modelo
df_kmeans = df_kmeans.rename(columns={'CLUSTER': 'CLUSTER_KMEANS'})
df_bkmeans = df_bkmeans.rename(columns={'CLUSTER': 'CLUSTER_BKMEANS'})
df_sc = df_sc.rename(columns={'CLUSTER': 'CLUSTER_SC'})


df_unificado = df_kmeans.merge(df_bkmeans, on='CUPS').merge(df_sc, on='CUPS')

df_unificado.to_csv('dataset/clusters_unificado.csv', index=False)

print(df_unificado.head())
